Lecture11/requirements.txt
python-dotenv
joblib
pandas
agno